requirements.txt
fastapi
uvicorn
pytesseract
pillow
opencv-python-headless
opencv-python
python-multipart
marker-pdf
flair
gensim
google-generativeai